Not exactly working, but some good notes so that I don't forget
Append can't return a nil because it returns a *Error, so the value will always by typed.
A new function that returns error would remove the ability to change the format of the message,
but otherwise would work well.

Maybe the only thing that needs to change is the usage. Build an array of errors first, then use multierror.Error
to call ErrorOrNil, or format it.

AppendError might work, allowing for formatting to be deferred, but the implementation needs more work.